cleanup by adding controller and scr folder
