--- routes/productRoutes.js
@@ -1,34 +1,0 @@
-import express from "express";
-import asyncHandler from "express-async-handler";
-import Product from "../models/productModel.js";
-
-const router = express.Router();
-
-// @desc     Fetch all Products
-// @route    Get /api/products
-// @access   Public
-router.get(
-  "/",
-  asyncHandler(async (req, res) => {
-    const products = await Product.find({});
-    res.json(products);
-  })
-);
-
-// @desc     Fetch single Products
-// @route    Get /api/products/:id
-// @access   Public
-router.get(
-  "/:id",
-  asyncHandler(async (req, res) => {
-    const product = await Product.findById(req.params.id);
-    if (product) {
-      res.json(product);
-    } else {
-      res.status(404);
-      throw new Error("Product not found");
-    }
-  })
-);
-
-export default router;
