+++ seeder.js
@@ -1,12 +1,12 @@
 import mongoose from "mongoose";
 import dotenv from "dotenv";
 import colors from "colors";
-import users from "./data/users.js";
-import products from "./data/products.js";
-import User from "./models/userModel.js";
-import Product from "./models/productModel.js";
-import Order from "./models/orderModel.js";
-import connectDB from "./config/db.js";
+import users from "./database/data/users.js";
+import products from "./database/data/products.js";
+import User from "./database/models/userModel.js";
+import Product from "./database/models/productModel.js";
+import Order from "./database/models/orderModel.js";
+import connectDB from "./src/config/db.js";
 
 dotenv.config();
