+++ server.js
@@ -1,9 +1,9 @@
 import express from "express";
 import dotenv from "dotenv";
 import colors from "colors";
-import connectDB from "./config/db.js";
-import productRoutes from "./routes/productRoutes.js";
-import { notFound, errorHandler } from "./middleware/errorMiddleware.js";
+import connectDB from "./src/config/db.js";
+import productRoutes from "./src/routes/productRoutes.js";
+import { notFound, errorHandler } from "./src/middlewares/errorMiddleware.js";
 
 const app = express();
 dotenv.config();
